Refactor: Повышение безопасности BLE и улучшение драйвера ШИМ
- Повышен уровень безопасности характеристик до ENC_NO_MITM.
- Улучшена функция pwm_handler_set_rgb: теперь она принимает стандартные значения RGB (0-255).

+++ pwm_handler.h
@@ -6,7 +6,7 @@
 // Инициализация модуля ШИМ
 void pwm_handler_init(const uint32_t *led_pins);
 
-// Установка цвета RGB (значения 0-1000)
-void pwm_handler_set_rgb(uint16_t r, uint16_t g, uint16_t b);
+// Установка цвета RGB (значения 0-255)
+void pwm_handler_set_rgb(uint8_t r, uint8_t g, uint8_t b);
 
 #endif
